requirements.txt
attrs==24.2.0
psycopg2-binary==2.9.9
certifi==2024.8.30
charset-normalizer==3.4.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
packaging==24.2
playwright==1.49.0
pyee==12.0.0
pyparsing==3.2.0
allure-pytest==2.13.5
pytest==7.4.3
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
wsproto==1.2.0
zstandard==0.23.0
